+++ HISTORY.rst
@@ -13,7 +13,7 @@ History
 - CHORE: more test coverage
 - CHORE: refactor endpoint urls to settings
 - CHORE: refactor message attributes from SNSMessage to base class
-
+- FIX: fix when message is not json loadable
 
 0.1.1 (2019-03-22)
 ++++++++++++++++++